BACKEND/requirements_clean.txt
Django==5.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.8
PyJWT==2.8.0
sqlparse==0.5.5
tzdata==2024.1
cloudinary==1.40.0
cloudinary-storage==0.4.0
python-dotenv==1.0.1
dj-database-url==2.2.0
